+++ src/Ext/AnimType/Hooks.cpp
@@ -12,3 +12,19 @@ DEFINE_HOOK(0x422CAB, AnimClass_DrawIt_XDrawOffset, 0x5)
 
 	return 0;
 }
+
+DEFINE_HOOK(0x423B95, AnimClass_AI_HideIfNoOre_Threshold, 0x8)
+{
+	GET(AnimClass* const, pThis, ESI);
+	GET(AnimTypeClass* const, pType, EDX);
+
+	if (pType->HideIfNoOre)
+	{
+		auto nThreshold = abs(AnimTypeExt::ExtMap.Find(pType)->HideIfNoOre_Threshold.Get());
+		auto pCell = pThis->GetCell();
+
+		pThis->Invisible = !pCell || pCell->GetContainedTiberiumValue() <= nThreshold;
+	}
+
+	return 0x423BBF;
+}